--- src/main/java/step2/domain/lotto/LottoStrategy.java
@@ -1,8 +1,0 @@
-package step2.domain.lotto;
-
-import java.util.List;
-
-public interface LottoStrategy {
-
-    List<Integer> createNumbers();
-}
