+++ src/main/java/step3/domain/lotto/LottoStrategy.java
@@ -1,0 +1,8 @@
+package step3.domain.lotto;
+
+import java.util.List;
+
+public interface LottoStrategy {
+
+    List<LottoNumber> createNumbers();
+}
