+++ src/main/java/step3/domain/lotto/Profit.java
@@ -1,6 +1,6 @@
-package step2.domain.lotto;
+package step3.domain.lotto;
 
-import step2.domain.money.Cache;
+import step3.domain.money.Cache;
 import java.util.Objects;
 
 public class Profit {
