+++ src/main/java/step3/domain/lotto/WinOfLotto.java
@@ -1,4 +1,4 @@
-package step2.domain.lotto;
+package step3.domain.lotto;
 
 import java.util.ArrayList;
 import java.util.List;
