+++ src/main/java/step3/util/LottoNumberGenerator.java
@@ -1,5 +1,6 @@
-package step2.util;
+package step3.util;
 
+import step3.domain.lotto.LottoNumber;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
@@ -10,23 +11,23 @@ public class LottoNumberGenerator {
     private LottoNumberGenerator() {
     }
 
-    private static final List<Integer> LOTTO_NUMBERS = new ArrayList<>();
+    private static final List<LottoNumber> LOTTO_NUMBERS = new ArrayList<>();
 
-    public static List<Integer> create() {
+    public static List<LottoNumber> create() {
         if (LOTTO_NUMBERS.isEmpty()) {
             createLottoNumbers();
         }
 
         Collections.shuffle(LOTTO_NUMBERS);
-        List<Integer> numbers = new ArrayList<>(LOTTO_NUMBERS.subList(0, 6));
-        numbers.sort(Comparator.naturalOrder());
+        List<LottoNumber> numbers = new ArrayList<>(LOTTO_NUMBERS.subList(0, 6));
+        numbers.sort(Comparator.comparing(LottoNumber::getNumber));
         return numbers;
 
     }
 
     private static void createLottoNumbers() {
         for (int i = 1; i < 46; i++) {
-            LOTTO_NUMBERS.add(i);
+            LOTTO_NUMBERS.add(new LottoNumber(i));
         }
     }
 }
