+++ src/test/java/step3/domain/LottoMachineTest.java
@@ -1,9 +1,9 @@
-package step2.domain;
+package step3.domain;
 
 import org.junit.jupiter.api.Test;
-import step2.domain.lotto.Lotto;
-import step2.domain.lotto.LottoMachine;
-import step2.domain.money.Cache;
+import step3.domain.lotto.Lotto;
+import step3.domain.lotto.LottoMachine;
+import step3.domain.money.Cache;
 import java.util.Arrays;
 import java.util.List;
